binder: MTK: Remove MTK's Debugging and Monitoring flags
Removed MTK's custom flags that are for debugging and monitoring purposes.
We dont really need them as we are not building an eng build(only active in it)
Removed flags are:
	1. BINDER_MONITOR
	2. MTK_BINDER_DEBUG
	3. RT_PRIO_INHERIT
	4. BINDER_PERF_EVAL
	5. MTK_BINDER_PAGE_USED_RECORD

This is also done to prevent conflicts when upstreaming from google's common kernel.

More flags might be removed later.

Signed-off-by: Sahil Soni <SscSPs@gmail.com>